Shift .sidebar-form back outside the role='main' div so that the width behaves better with the transition.
Pull .sidebar-form out to a directive so that its easy to include on all views (needs access to the lists controller)

Add addTodo() method on the TodoList controller so that the form submit triggers an add.  This may be breaking other, working behavior, but I couldn't figure out what was _supposed_ to be happening with the todo controller.